parameterization

+++ src/helpers.jl
@@ -443,11 +443,11 @@ function make_optim_pars(optim_pars; solver = "newtontrust")
                 rho_upper = 0.75)
         else
             optim_pars = (
-                initial_delta = haskey(optim_pars, :initial_delta) ? optim_pars.initial_delta : 1.0,
-                delta_hat = haskey(optim_pars, :delta_hat) ? optim_pars.delta_hat : 5.0,
-                eta = haskey(optim_pars, :eta) ? optim_pars.eta : 0.1,
-                rho_lower = haskey(optim_pars, :rho_lower) ? optim_pars.rho_lower : 0.25,
-                rho_upper = haskey(optim_pars, :rho_upper) ? optim_pars.rho_upper : 0.75)
+                initial_delta = haskey(optim_pars, :initial_delta) ? Float64(optim_pars.initial_delta) : 1.0,
+                delta_hat = haskey(optim_pars, :delta_hat) ? Float64(optim_pars.delta_hat) : 5.0,
+                eta = haskey(optim_pars, :eta) ? Float64(optim_pars.eta) : 0.1,
+                rho_lower = haskey(optim_pars, :rho_lower) ? Float64(optim_pars.rho_lower) : 0.25,
+                rho_upper = haskey(optim_pars, :rho_upper) ? Float64(optim_pars.rho_upper) : 0.75)
         end
     end
     return optim_pars
